:white_check_mark: adds test
:bug: adds back tests
